Lab 04 not complete; red description
Can't remove the lag when being killed by orc
Also can't find NGUI ofr Lab 05
Pls help

+++ Assets/Content/Scripts/Scene/LevelController.cs
@@ -24,10 +24,10 @@ public class LevelController : MonoBehaviour
         StartCoroutine(waitForDeath(1, rabbit));
     }
 
-    IEnumerator waitForDeath(int sec, HeroController rabbit)
+    IEnumerator waitForDeath(float sec, HeroController rabbit)
     {
         Animator animator = rabbit.GetComponent<Animator>();
-        animator.SetBool("die", true);
+        animator.SetBool("die",true);
         // Debug.LogWarning("Starting");
         yield return new WaitForSeconds(sec);
         // Debug.LogWarning("End");
